wip(UIKIT-1430,AutoSaveIndicator): Правки по PR
