+++ packages/components/src/AutoSaveIndicator/DefaultState/index.ts
@@ -1,0 +1,1 @@
+export * from './DefaultState';
