+++ packages/components/src/AutoSaveIndicator/DefaultState/styles.ts
@@ -1,0 +1,17 @@
+import { tooltipClasses } from '@mui/material';
+
+import { styled } from '../../styles/styled';
+import { Tooltip } from '../../Tooltip';
+import { Typography } from '../../Typography';
+
+export const StyledTooltip = styled(Tooltip)`
+  & .${tooltipClasses.tooltip} {
+    max-width: none;
+  }
+`;
+
+export const StyledTypography = styled(Typography)`
+  align-self: flex-end;
+
+  padding-right: ${({ theme }) => theme.spacing(2)};
+`;
