+++ packages/components/src/AutoSaveIndicator/SuccessState/SuccessState.tsx
@@ -1,0 +1,12 @@
+import { SUCCESS_STATE } from '../constants';
+
+import { StyledTypography, SuccessIcon } from './styles';
+
+export const SuccessState = () => {
+  return (
+    <>
+      <StyledTypography variant="caption">{SUCCESS_STATE}</StyledTypography>
+      <SuccessIcon />
+    </>
+  );
+};
