+++ packages/icons/generated-themed-icons/AutosaveLoaderFillMd.tsx
@@ -15,7 +15,6 @@ const AutosaveLoaderFillMd: React.FunctionComponent<SvgIconProps> = ({
       fill-rule="evenodd"
       clip-rule="evenodd"
       d="M7 20H17.5H18V19.9776C20.8032 19.725 23 17.369 23 14.5C23 11.9332 21.2417 9.77712 18.8639 9.17043C18.2503 6.21826 15.6342 4 12.5 4C9.78783 4 7.46364 5.6611 6.48927 8.02143C3.41461 8.28061 1 10.8583 1 14C1 17.3137 3.68629 20 7 20ZM12.6928 10.5994C12.2871 10.4821 11.8583 10.4694 11.4464 10.5626C11.0345 10.6558 10.653 10.8518 10.3373 11.1324C10.0217 11.4129 9.78222 11.7689 9.64134 12.167L9.47454 12.6383L8.53183 12.3047L8.69862 11.8334C8.89585 11.276 9.23106 10.7777 9.67297 10.3849C10.1149 9.99217 10.6491 9.71774 11.2257 9.58727C11.8024 9.45679 12.4027 9.47452 12.9706 9.6388C13.536 9.80233 14.0509 10.1058 14.4679 10.521L15.1667 11.1776V10.3335C15.1667 10.0574 15.3905 9.8335 15.6667 9.8335C15.9428 9.8335 16.1667 10.0574 16.1667 10.3335V12.3335V12.8335H15.6717H15.6614H13.6667C13.3905 12.8335 13.1667 12.6096 13.1667 12.3335C13.1667 12.0574 13.3905 11.8335 13.6667 11.8335H14.4043L13.7776 11.2446L13.7719 11.2392L13.7663 11.2337C13.4678 10.935 13.0985 10.7168 12.6928 10.5994ZM7.83331 13.1665V13.6665V13.667V15.6665C7.83331 15.9426 8.05717 16.1665 8.33331 16.1665C8.60946 16.1665 8.83331 15.9426 8.83331 15.6665V14.8228L9.53209 15.4794C9.94904 15.8946 10.464 16.198 11.0293 16.3616C11.5973 16.5259 12.1976 16.5436 12.7742 16.4131C13.3509 16.2826 13.8851 16.0082 14.327 15.6154C14.7689 15.2227 15.1041 14.7243 15.3013 14.167L15.4681 13.6956L14.5254 13.362L14.3586 13.8334C14.2177 14.2315 13.9783 14.5874 13.6627 14.868C13.347 15.1485 12.9654 15.3446 12.5535 15.4378C12.1416 15.531 11.7129 15.5183 11.3072 15.401C10.9015 15.2836 10.5322 15.0654 10.2336 14.7667L10.2281 14.7612L10.2224 14.7558L9.59522 14.1665H10.3333C10.6095 14.1665 10.8333 13.9426 10.8333 13.6665C10.8333 13.3904 10.6095 13.1665 10.3333 13.1665H8.33331H7.83331Z"
-      fill="#1D3F66"
     />
   </SvgIcon>
 );
